rename path2d to path
