+++ lib/classes/context.js
@@ -9,7 +9,7 @@ const {RustClass, core, wrap, inspect, REPR} = require('./neon'),
       {fromSkMatrix, toSkMatrix} = require('./geometry'),
       {Image, ImageData} = require('./imagery'),
       {TextMetrics} = require('./typography'),
-      {Path2D} = require('./path2d'),
+      {Path2D} = require('./path'),
       css = require('./css')
 
 const toString = val => typeof val=='string' ? val : new String(val).toString()
