+++ lib/classes/path.js
