+++ lib/index.js
@@ -10,7 +10,7 @@ const {Canvas, CanvasGradient, CanvasPattern, CanvasTexture} = require('./classe
       {TextMetrics, FontLibrary} = require('./classes/typography'),
       {CanvasRenderingContext2D} = require('./classes/context'),
       {App, Window} = require('./classes/gui'),
-      {Path2D} = require('./classes/path2d')
+      {Path2D} = require('./classes/path')
 
 module.exports = {
   Canvas, CanvasGradient, CanvasPattern, CanvasTexture,
